+++ Gemfile
@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
+gem 'arkana'
 gem "cocoapods"
 gem "cocoapods-clean"
 gem "cocoapods-keys"
-
